implemented the functionalities for user administration and token authentication

+++ controllers/blogs.js
@@ -1,8 +1,10 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
   
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
     response.json(blogs)
   })
 
@@ -17,14 +19,43 @@ blogsRouter.get('/:id', async (request, response, next) => {
   })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+
+
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  const blog = await Blog.findById(request.params.id)
+
+  if(blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(400).end()
+  }
    
-      const blog = await Blog.findByIdAndRemove(request.params.id)
-      response.status(204).end()
-    
   })
 
+const getTokenFrom = request => {
+    const authorization = request.get('authorization')
+    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+      return authorization.substring(7)
+    }
+    return null
+  }
+
 blogsRouter.post('/', async (request, response) => {
-    const body = request.body
+  const body = request.body
+
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
   
     if (!body.title || !body.author || !body.url) {
       return response.status(400).json({ error: "The fields 'title', 'author' and 'link' must be provided" })
@@ -35,11 +66,14 @@ blogsRouter.post('/', async (request, response) => {
       author: body.author,
       url: body.url,
       likes: body.likes,
-      status: body.status
+      status: body.status,
+      user: user._id
     })
   
   
       const savedBlog = await blog.save()
+      user.blogs = user.blogs.concat(savedBlog._id)
+      await user.save()
       response.status(201).json(savedBlog)
    
   })
